refactor(rust): split monolithic refactor.py into package
Split the 1,146-line refactor.py into a refactor/ Python package with
7 focused modules (all under 250 lines):

- __init__.py / __main__.py — package entry + command dispatch
- types.py — shared data structures
- parsing.py — parse_items command
- tests.py — find_related_tests command
- imports.py — resolve_imports command
- visibility.py — adjust_visibility command
- rewrite.py — rewrite_caller_imports command
- struct_fields.py — struct field extraction helpers

Updated refactor.sh wrapper to use PYTHONPATH + python3 -m refactor.
All 6 commands verified working end-to-end.
